add conftest, rearrange tests
